requirements.txt
streamlit
pandas
google-generativeai
python-docx
requests
streamlit-oauth
auth0-python
psycopg2-binary